LayoutTests/fast/css-grid-layout/grid-auto-columns-rows-get-set-expected.txt

Test that setting and getting grid-auto-columns and grid-auto-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-auto-columns and -webkit-grid-auto-rows set through CSS
PASS getComputedStyle(gridAutoFixedFixed, '').getPropertyValue('-webkit-grid-auto-rows') is '30px'
PASS getComputedStyle(gridAutoFixedFixed, '').getPropertyValue('-webkit-grid-auto-columns') is '50px'
PASS getComputedStyle(gridAutoMinMax, '').getPropertyValue('-webkit-grid-auto-rows') is 'minmax(10%, 15px)'
PASS getComputedStyle(gridAutoMinMax, '').getPropertyValue('-webkit-grid-auto-columns') is 'minmax(30%, 100px)'
PASS getComputedStyle(gridAutoMinMaxContent, '').getPropertyValue('-webkit-grid-auto-rows') is '-webkit-min-content'
PASS getComputedStyle(gridAutoMinMaxContent, '').getPropertyValue('-webkit-grid-auto-columns') is '-webkit-max-content'

Test the initial value
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'

Test getting and setting -webkit-grid-auto-columns and -webkit-grid-auto-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is '180px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is '660px'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'minmax(-webkit-min-content, 48px)'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'minmax(80px, -webkit-min-content)'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'minmax(-webkit-min-content, -webkit-max-content)'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'minmax(-webkit-max-content, -webkit-min-content)'

Test setting grid-auto-columns and grid-auto-rows to bad minmax value through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-auto-rows') is 'auto'
PASS successfullyParsed is true

TEST COMPLETE
